+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Run `preToolCall` hook before user approval if any. #170
+
 ## 0.72.1
 
 - Only include `parallel_tool_calls` to openai-responses and openai-chat if true. #169